+++ README.md
@@ -1,5 +1,5 @@
-# 📂 DBAssets
-DBAssets is a tool to quickly store and retrieve compromised credentials and hosts; the goal is to ease the management of credentials and hosts during a penetration testing engagement or a CTF.
+# 📂 Exegol-history
+Exegol-history is a tool to quickly store and retrieve compromised credentials and hosts; the goal is to ease the management of credentials and hosts during a penetration testing engagement or a CTF.
 
 Once an asset is selected from the TUI, the information can be accessed through environment variables and doesn't need to be typed over and over.
 
@@ -7,13 +7,13 @@ For CLI usage, see [Wiki](WIKI.md)
 
 ## ⚙️ Install
 ```
-pipx install git+https://github.com/lap1nou/db-assets
+pipx install git+https://github.com/lap1nou/exegol-history
 ```
 
 You can then add these two shell aliases in your shell configuration file (`.bashrc`, `.zshrc`, ...):
 ```bash
-alias dbcreds='(){
-	CREDS=$(dbassets tui creds);
+alias exegolcreds='(){
+	CREDS=$(exegol-history tui creds);
 	export USER=$( echo "$CREDS" | sed "1q;d");
 	export PASSWORD=$( echo "$CREDS" | sed "2q;d");
 	export NT_HASH=$( echo "$CREDS" | sed "3q;d");
@@ -22,8 +22,8 @@ alias dbcreds='(){
 ```
 
 ```bash
-alias dbhosts='(){
-	HOSTS=$(dbassets tui hosts);
+alias exegolhosts='(){
+	HOSTS=$(exegol-history tui hosts);
 	export IP=$( echo "$HOSTS" | sed "1q;d");
 	export TARGET=$( echo "$HOSTS" | sed "1q;d");
 	export HOSTNAME=$( echo "$HOSTS" | sed "2q;d");
@@ -36,7 +36,7 @@ To see which user you are currently using, you can add the `USER` environnment v
 variable = "USER"
 default = ''
 style = "fg:bold red bg:#477069"
-format = '[  $env_value ]($style)'
+format = '[  $env_value ]($style)'
 ```
 
 ## ✨ Features
@@ -52,7 +52,7 @@ The TUI (Terminal User Interface):
 
 ![](screenshots/screenshot03.png)
 
-A typical workflow with `DBAssets`:
+A typical workflow with `Exegol-history`:
 ![](screenshots/screenshot04.png)
 
 ## 📝 Configuration
@@ -77,7 +77,7 @@ edit_host = "f5"
 quit = "ctrl+c"
 ```
 
-**The configuration file must be in the home folder, in a `.dbassets` folder.**
+**The configuration file must be in the home folder, in a `.exegol-history` folder.**
 
 ## 🛠️ Development
