--- db-assets.py
@@ -1,4 +1,0 @@
-from dbassets.__main__ import main
-
-if __name__ == "__main__":
-    main()
