+++ exegol_history/tests/artifacts/pypykatz.txt
