+++ exegol_history/tests/cli_tests/conftest.py
@@ -1,12 +1,12 @@
 import os
 import pytest
 
-from dbassets.__main__ import setup
+from exegol_history.__main__ import setup
 from pykeepass import PyKeePass
 
 TEST_DB_NAME = "test.kdbx"
 TEST_KEY_NAME = "test.key"
-TEST_ARTIFACTS_PATH = os.path.join("dbassets", "tests", "artifacts")
+TEST_ARTIFACTS_PATH = os.path.join("exegol_history", "tests", "artifacts")
 
 TEST_DB_PATH = os.path.join(TEST_ARTIFACTS_PATH, TEST_DB_NAME)
 TEST_KEY_PATH = os.path.join(TEST_ARTIFACTS_PATH, TEST_KEY_NAME)
