+++ exegol_history/tests/cli_tests/test_delete_credential.py
@@ -1,6 +1,6 @@
 import pytest
 
-from dbassets.db_api.creds import add_credential, get_credentials, delete_credential
+from exegol_history.db_api.creds import add_credential, get_credentials, delete_credential
 from common_cli import (
     USERNAME_TEST_VALUE,
     PASSWORD_TEST_VALUE,
