+++ exegol_history/tests/cli_tests/test_delete_host.py
@@ -1,6 +1,6 @@
 import pytest
 
-from dbassets.db_api.hosts import add_host, get_hosts, delete_host
+from exegol_history.db_api.hosts import add_host, get_hosts, delete_host
 from common_cli import IP_TEST_VALUE, HOSTNAME_TEST_VALUE, ROLE_TEST_VALUE
 from pykeepass import PyKeePass
