+++ exegol_history/tests/cli_tests/test_edit_credential.py
@@ -1,4 +1,4 @@
-from dbassets.db_api.creds import add_credential, get_credentials, edit_credential
+from exegol_history.db_api.creds import add_credential, get_credentials, edit_credential
 from common_cli import (
     USERNAME_TEST_VALUE,
     PASSWORD_TEST_VALUE,
