+++ exegol_history/tests/tui_tests/test_add_credential_tui.py
@@ -1,8 +1,8 @@
 import pytest
 import os
 
-from dbassets.tui.db_creds.db_creds import DbCredsApp
-from dbassets.db_api.creds import get_credentials
+from exegol_history.tui.db_creds.db_creds import DbCredsApp
+from exegol_history.db_api.creds import get_credentials
 from common_tui import (
     USERNAME_TEST_VALUE,
     PASSWORD_TEST_VALUE,
