+++ exegol_history/tui/css/delete_object.tcss
